feat(doctor): report the privileged helper's org signature status
Adds a `veld doctor` check (privileged mode only) that verifies the on-disk
helper's detached .sig against the embedded org ed25519 key, against the
SERVICE-pinned binary the helper's gate actually verifies. A signature mismatch
means the next `veld update` will sit on the old version, so this turns that
from a mysterious stuck update into a visible, actionable row with the exact
recovery command (redeploy, or sudo launchctl/systemctl to force the binary).

To share one key and verify path between the helper's gate and the doctor,
centralize the org pubkey, bounded .sig read, and verification in
`veld_core::signing`; `veld-helper/src/signing.rs` re-exports it. A duplicated
constant would drift on key rotation, and two verify paths would disagree about
the one thing a security gate must not guess.
