+++ crates/veld-core/Cargo.toml
@@ -21,6 +21,7 @@ rusqlite = { workspace = true }
 nix = { workspace = true }
 reqwest = { workspace = true }
 sha2 = { workspace = true }
+ed25519-dalek = { workspace = true }
 base64 = { workspace = true }
 petname = { workspace = true }
 indexmap = { workspace = true }
