+++ crates/veld/src/commands/doctor.rs
@@ -471,6 +471,34 @@ impl Diagnostics {
                 .await;
         }
 
+        // The privileged helper's fail-closed signing gate (#261) refuses to
+        // relaunch onto an on-disk binary without a valid org .sig — so a
+        // signature mismatch means the next `veld update` will sit on the old
+        // version. Check the SERVICE-pinned binary (the one the gate actually
+        // verifies via `current_exe()`), not a lib-dir guess: on a machine
+        // carrying both ~/.local and /usr/local, the guess would accuse a file
+        // the helper never runs.
+        if mode.as_deref() == Some("privileged") {
+            let path = veld_core::setup::privileged_helper_program()
+                .await
+                .unwrap_or_else(|| std::path::PathBuf::from(self.helper_path.clone()));
+            match veld_core::signing::relaunch_guard(&path) {
+                None => self.checks.push(Check {
+                    pass: true,
+                    label: format!("Helper binary signature OK ({})", tilde_path(&path)),
+                }),
+                Some(reason) => self.checks.push(Check {
+                    pass: false,
+                    label: format!(
+                        "{reason} — run `veld update` or `veld setup privileged` to \
+                         redeploy it; to force the installed binary now: \
+                         `sudo launchctl kill TERM system/dev.veld.helper` \
+                         (Linux: `sudo systemctl restart veld-helper`)"
+                    ),
+                }),
+            }
+        }
+
         // Determine HTTPS port for later checks. When the helper is down we
         // can't ask it, so fall back based on the configured mode — privileged
         // serves on 443; probing 18443 there checks a port nothing should be
